Fixed audio focus handling
