+++ ultrasonic/src/main/kotlin/org/moire/ultrasonic/di/MediaPlayerModule.kt
@@ -10,6 +10,7 @@ import org.moire.ultrasonic.service.LocalMediaPlayer
 import org.moire.ultrasonic.service.MediaPlayerController
 import org.moire.ultrasonic.service.MediaPlayerControllerImpl
 import org.moire.ultrasonic.service.MediaPlayerLifecycleSupport
+import org.moire.ultrasonic.service.AudioFocusHandler
 import org.moire.ultrasonic.util.ShufflePlayBuffer
 
 val mediaPlayerModule = module {
@@ -23,7 +24,8 @@ val mediaPlayerModule = module {
     single { ExternalStorageMonitor(androidContext()) }
     single { ShufflePlayBuffer(androidContext()) }
     single { Downloader(androidContext(), get(), get(), get()) }
-    single { LocalMediaPlayer(androidContext()) }
+    single { LocalMediaPlayer(get(), androidContext()) }
+    single { AudioFocusHandler(get()) }
 
     // TODO Ideally this can be cleaned up when all circular references are removed.
     single { MediaPlayerControllerImpl(androidContext(), get(), get(), get(), get(), get()) }
